QueueLocking: handle complex keys
We support complex keys because Avro schema supports them (and we've got
at least one topic that uses them). This support is reflected in the
fact that we use jsonb columns to store them in PostgreSQL.

When constructing lock keys, make sure to do so in a way that produces
equal results given equivalent (or identical) complex keys.