xid
Draft an interesting solution to the message ordering race-condition
that can arise during concurrent inserts.

https://event-driven.io/en/ordering_in_postgres_outbox/

This is an interesting and robust solution, although it has a few
drawbacks. First, it adds complexity, though not all that much. Second,
it's only supported on PostgreSQL 13 and newer. Finally, it can
introduce noticeable message-processing delays if run in an app that has
any long-running transactions, regardless of whether those transactions
involve messages.

Ultimately, the message-processing delay was the deal-breaker, because
we can't anticipate what kind of transaction activity will be running in
the downstream app that choses to use our engine. It could easily have
long-running transactions that have nothing to do with the inbox/outbox,
and these would end up delaying message-processing unnecessarily.