Ruby 3.2
Add support for Ruby 3.2 and update development/testing environment to
use that version.

NOTE: The Ruby 3.2 Docker image we use in CI doesn't include the libyaml
header files needed by the psych gem (a dependenciy of Rails). That has
been fixed in the Ruby 3.3 image, and I'll be upgrading everything to
3.3 in the next commit.