fix(rbi): ensure ActiveRecord::Transaction is defined
Fix coil's RBI file so that type-checking works regardless of which
Rails version the host application is using.

In 4078d45 we added support for Rails versions 7.2 and newer, in which
an `ActiveRecord::Transaction` object gets passed to the provided block
when calling `ActiveRecord::Base.transaction()`. The type-signature we
provided with that change was intended to support both old and new Rails
versions by using `T.any`:

    blk: T.any(
      # ActiveRecord versions 7.2 and newer yield a transaction as the block arg.
      T.proc.params(arg0: ActiveRecord::Transaction).returns(T.type_parameter(:P)),
      # Earlier versions yield to the block with no args.
      T.proc.returns(T.type_parameter(:P))
    )

But the `ActiveRecord::Transaction` class did not even exist prior to
version 7.2, causing type-checking to fail.

Fix that by conditionally defining `ActiveRecord::Transaction` in the
RBI file.