fix concurrency test
Fix tests around transactional message concurrency control. The
race-condition these tests are concerned with is:

    Client A | BEGIN..INSERT..............COMMIT
    Client B |    BEGIN..INSERT..COMMIT

Transactional messages have an `around_save` hook that obtains an
advisory lock, to prevent the above scenario if Clients A and B are
inserting the same type of message with the same key.

The tests were set up as if Client A obtained a lock as soon as it began
a database transaction, but that's not accurate. It only obtains a lock
upon calling `save!`. (It then holds the lock until the end of the
transaction.) This incorrect resulted flaky tests.